+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 Note: Numbers like (#123) point to closed Pull Requests on the fractal-vizarr-viewer repository.
 
+# Unreleased
+
+* Added support for S3 paths (\#82);
+* Fixed handling of invalid range requests (\#82);
+* Dropped Node 18 support, now requiring Node 20 (\#82);
+
 # 0.4.3
 
 * Update `User` type (\#74);